+++ lib/easy_dash_logger.dart
@@ -2,8 +2,8 @@ library easy_dash_logger;
 
 import 'dart:developer' as dev;
 
-import 'src/platform_specific_logger/flutter_logger.dart'
-    if (dart.library.io) 'src/platform_specific_logger/dart_logger.dart';
+import 'src/platform_specific_logger/is_dart.dart'
+    if (dart.library.ui) 'src/platform_specific_logger/is_flutter.dart';
 
 part 'src/enums.dart';
 part 'src/logger.dart';