+++ lib/src/platform_specific_logger/dart_logger.dart
@@ -1,9 +1,0 @@
-bool debugMode = false;
-
-bool isFlutterEnvironment() {
-  return false;
-}
-
-bool isDebugMode() {
-  return debugMode;
-}