+++ lib/src/platform_specific_logger/is_dart.dart
@@ -1,0 +1,2 @@
+// is_dart.dart
+bool isFlutter() => false;