+++ lib/src/platform_specific_logger/is_flutter.dart
@@ -1,0 +1,2 @@
+// is_flutter.dart
+bool isFlutter() => true;