+++ test/easy_dash_logger_test.dart
@@ -1,68 +1,48 @@
-import 'dart:async';
-
 import 'package:easy_dash_logger/easy_dash_logger.dart';
-import 'package:easy_dash_logger/src/platform_specific_logger/dart_logger.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-// Override the isDebugMode function for testing
-bool Function() isDebugModeOverride = isDebugMode;
-
-bool isDebugMode() {
-  return isDebugModeOverride();
-}
+import 'package:easy_dash_logger/src/platform_specific_logger/is_dart.dart'
+    if (dart.library.ui) 'package:easy_dash_logger/src/platform_specific_logger/is_flutter.dart';
+import 'package:test/test.dart';
 
 void main() {
-  group('Logger Tests', () {
-    test('Prints log in debug mode', () {
-      const message = 'Test log message in debug mode';
-      const name = 'DEBUG_LOG';
-      const color = ANSICOLOR.green;
-      const bgColor = ANSICOLOR.black;
-
-      // Set debugMode to true
-      debugMode = true;
-
-      // Capture the log output
-      final log = <String>[];
-      final spec = ZoneSpecification(print: (_, __, ___, String msg) {
-        log.add(msg);
-      });
-
-      // Run the test in a zone with the print interceptor
-      runZoned(() {
-        message.printLog(name: name, color: color, bgColor: bgColor);
-      }, zoneSpecification: spec);
-
-      // Verify the log output
-      print('Log output: $log');
-      expect(log, isNotEmpty);
-      expect(log.any((msg) => msg.contains(message)), isTrue);
+  group('EasyDashLogger Tests', () {
+    test('Default log message', () {
+      expect(() => 'This is a default log message'.printLog(), returnsNormally);
     });
 
-    test('Prints log in release mode', () {
-      const message = 'Test log message in release mode';
-      const name = 'RELEASE_LOG';
-      const color = ANSICOLOR.red;
-      const bgColor = ANSICOLOR.blue;
+    test('Custom log message', () {
+      expect(
+          () => 'This is a custom log message'.printLog(
+                name: 'CUSTOM_LOG',
+                color: ANSICOLOR.green,
+              ),
+          returnsNormally);
+    });
 
-      // Set debugMode to false
-      debugMode = false;
+    test('Log message with background color', () {
+      expect(
+          () => 'This is a log message with background color'.printLog(
+                color: ANSICOLOR.red,
+                bgColor: ANSICOLOR.lightYellow,
+              ),
+          returnsNormally);
+    });
 
-      // Capture the print output
-      final log = <String>[];
-      final spec = ZoneSpecification(print: (_, __, ___, String msg) {
-        log.add(msg);
-      });
-
-      // Run the test in a zone with the print interceptor
-      runZoned(() {
-        message.printLog(name: name, color: color, bgColor: bgColor);
-      }, zoneSpecification: spec);
-
-      // Verify the log output
-      print('Log output: $log');
-      expect(log, isNotEmpty);
-      expect(log.any((msg) => msg.contains(message)), isTrue);
+    test('Log message in Flutter environment', () {
+      if (isFlutter()) {
+        expect(
+            () => 'This is a log message in Flutter'.printLog(
+                  name: 'FLUTTER_LOG',
+                  color: ANSICOLOR.blue,
+                ),
+            returnsNormally);
+      } else {
+        expect(
+            () => 'This is a log message in Dart'.printLog(
+                  name: 'DART_LOG',
+                  color: ANSICOLOR.purple,
+                ),
+            returnsNormally);
+      }
     });
   });
 }